feat:6424

+++ WeeklyContest/2712.go
@@ -5,7 +5,7 @@ func minimumCost(s string) int64 {
     var sum int64
     for i := 1; i < n; i++ {
         if s[i] != s[i-1] {
-            sum += min(i, n-i)
+            sum += int64(min(i, n-i))
         }
     }
     return sum
